// services/core-service/src/seeder/old-schema.ts
import { RoomType } from 'shared-types';
import { BusinessCategoryDocument } from '../schemas/business-category.schema';
import { LanguageDocument } from '../schemas/language.schema';
import { MeetingInstanceDocument } from '../schemas/meeting-instance.schema';
import { PreviewImageDocument } from '../schemas/preview-image.schema';
import { SocialLinkDocument } from '../schemas/social-link.schema';
import { UserDocument } from '../schemas/user.schema';

export interface OldUserTemplate {
  _id: any;
  templateId: number;
  usedAt: number;
  timesUsed: number;
  url: string;
  draftUrl: string;
  name: string;
  maxParticipants: number;
  businessCategories: BusinessCategoryDocument[];
  description: string;
  shortDescription: string;
  previewUrls: PreviewImageDocument[];
  draftPreviewUrls: PreviewImageDocument[];
  type: string;
  priceInCents: number;
  fullName: string;
  companyName: string;
  position: string;
  contactEmail: string;
  languages: LanguageDocument[];
  socials: SocialLinkDocument[];
  user: UserDocument;
  meetingInstance: MeetingInstanceDocument;
  signBoard: string;
  customLink: string;
  templateCurrency: string;
  templatePrice: number;
  paywallCurrency: string;
  paywallPrice: number;
  isMonetizationEnabled: boolean;
  isAudioAvailable: boolean;
  usersPosition: { bottom: number; left: number }[];
  usersSize: number[];
  indexUsers: string[];
  links: { item: string; position: { top: number; left: number } }[];
  author: UserDocument;
  draft: boolean;
  isPublic: boolean;
  templateType: 'video' | 'image';
  isAcceptNoLogin: boolean;
  roomType: RoomType;
  subdomain: string;
}